// src/app/models/clients/contacts.model.ts
export class ContactsModel {
  leadId: number;
  firstName: string;
  surName: string;
  otherName: string;
  dateOfBirth: string;
  gender: string;
  maritalStatus: string;
  maidenName: string;
  occupation: string;
  personalPhoneNumber: number;
  officialPhoneNumber: number;
  email: string;
  otherEmail: string;
  postalAddress: string;
  fax: string;
  zip: string;
  stateOfOrigin: string;
  nextOfKin: string;
  nextOfKinRelationship: string;
  profileImage: string;
  otherInfo: string;
  title: string;
}